backend/requirements.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
django-filter>=23.5,<25.0
django-auditlog>=2.3,<3.0
dj-database-url>=2.1,<3.0

# Async tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# Web server
gunicorn>=21.0,<23.0

# PDF generation
WeasyPrint>=60.0,<62.0
Jinja2>=3.1,<4.0

# AWS
boto3>=1.28,<2.0

# Security
cryptography>=41.0,<43.0
requests>=2.31,<3.0

# Testing
pytest>=7.4,<9.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<6.0
factory-boy>=3.3,<4.0
responses>=0.23,<1.0
freezegun>=1.2,<2.0
